+++ ReleaseNotes.md
@@ -16,6 +16,7 @@
 - Shake is now linked to target (through a Shaker behaviour), not to action itself - now we can destroy the object with the action when it's triggered
 - Shake returns object to initial state
 - Added warning when a button name is not defined (or doesn't exist)
+- Added warning when destroy action is not the last action in the list
 
 ## V1.11.1